requirements.txt
asgiref==3.8.1
Django==5.1.6
Jinja2==3.1.5
libvirt-python==11.0.0
MarkupSafe==3.0.2
sqlparse==0.5.3
